Fix styles for Tic-Tac-Toe game and refactor agent
- Refactored the `calculateWinner` function to a shared utility to avoid code duplication and ensure consistency between the game logic and the AI agent.
- Updated the AI agent in `agent.tsx` to use the shared `calculateWinner` function.
- Corrected the CSS class names in `ttt.css` to match the class names used in the `Game2.tsx` component, so the styles are applied correctly.
