+++ src/components/games/Game2.tsx
@@ -2,6 +2,7 @@ import React, { useState, useEffect } from 'react';
 import { useAuth } from '../../contexts/AuthContext';
 import Board from './board';
 import { getBestMove } from './agent';
+import { calculateWinner } from './utils';
 import './ttt.css';
 
 const Game2: React.FC = () => {
@@ -64,7 +65,7 @@ const Game2: React.FC = () => {
     <div className="game-container">
       <div className="game-header">
         <h1>Tic Tac Toe</h1>
-        <p>{getStatus()}</p>
+        <p className="status">{getStatus()}</p>
       </div>
       <div className="game-board">
         <Board board={board} onClick={handleClick} />
@@ -79,29 +80,6 @@ const Game2: React.FC = () => {
       </div>
     </div>
   );
-};
-
-// winning situations
-const calculateWinner = (board: (string | null)[]) => {
-  const winningLines = [
-    [0, 1, 2], // rows
-    [3, 4, 5],
-    [6, 7, 8],
-    [0, 3, 6], // columns
-    [1, 4, 7],
-    [2, 5, 8],
-    [0, 4, 8], // diagonals
-    [2, 4, 6],
-  ];
-
-  // loop to go through all case
-  for (let line of winningLines) {
-    const [a, b, c] = line;
-    if (board[a] && board[a] === board[b] && board[a] === board[c]) {
-      return board[a];
-    }
-  }
-  return null;
 };
 
 export default Game2;
