+++ src/components/games/utils.ts
@@ -1,0 +1,22 @@
+// winning situations
+export const calculateWinner = (board: (string | null)[]) => {
+  const winningLines = [
+    [0, 1, 2], // rows
+    [3, 4, 5],
+    [6, 7, 8],
+    [0, 3, 6], // columns
+    [1, 4, 7],
+    [2, 5, 8],
+    [0, 4, 8], // diagonals
+    [2, 4, 6],
+  ];
+
+  // loop to go through all case
+  for (let line of winningLines) {
+    const [a, b, c] = line;
+    if (board[a] && board[a] === board[b] && board[a] === board[c]) {
+      return board[a];
+    }
+  }
+  return null;
+};
